Updating loading, consolidating source code, notebook for downloading
