+++ src/data/fetch.py
@@ -1,7 +1,59 @@
 import numpy as np
 import requests
-import io
+from io import BytesIO
 from imageio import imread
+
+from .gee_utils import (get_sentinel2_download_url, get_landtrendr_download_url,
+                        read_gee_url)
+
+
+def dem_from_tnm(bbox, width, height, inSR=3857, **kwargs):
+    """
+    Retrieves a Digital Elevation Model (DEM) image from The National Map (TNM)
+    web service.
+
+    Parameters
+    ----------
+    bbox : list-like
+      list of bounding box coordinates (minx, miny, maxx, maxy)
+    width, height : int
+      desired width and height of returned image
+    inSR : int
+      spatial reference for bounding box, such as an EPSG code (e.g., 4326)
+
+    Returns
+    -------
+    dem : arr
+      DEM image as array
+    """
+    BASE_URL = ''.join([
+        'https://elevation.nationalmap.gov/arcgis/rest/',
+        'services/3DEPElevation/ImageServer/exportImage?'
+    ])
+
+    params = dict(bbox=','.join([str(x) for x in bbox]),
+                  bboxSR=inSR,
+                  size=f'{width},{height}',
+                  imageSR=inSR,
+                  time=None,
+                  format='tiff',
+                  pixelType='F32',
+                  noData=None,
+                  noDataInterpretation='esriNoDataMatchAny',
+                  interpolation='+RSP_BilinearInterpolation',
+                  compression=None,
+                  compressionQuality=None,
+                  bandIds=None,
+                  mosaicRule=None,
+                  renderingRule=None,
+                  f='image')
+    for key, value in kwargs.items():
+        params.update({key: value})
+
+    r = requests.get(BASE_URL, params=params)
+    dem = imread(BytesIO(r.content))
+
+    return dem
 
 
 def nlcd_from_mrlc(bbox, width, height, layer, inSR=4326, nlcd=True, **kwargs):
@@ -42,7 +94,7 @@ def nlcd_from_mrlc(bbox, width, height, layer, inSR=4326, nlcd=True, **kwargs):
         params.update({key: value})
 
     r = requests.get(BASE_URL, params=params)
-    img = imread(io.BytesIO(r.content), format='tiff')
+    img = imread(BytesIO(r.content), format='tiff')
 
     if nlcd:
         MAPPING = {
@@ -120,3 +172,70 @@ def colorize_nlcd(img):
         land_color[mask] = COLOR_MAP[cov]
 
     return land_color
+
+
+def landtrendr_from_gee(bbox, year, epsg, scale=30):
+    """Fetches an 8-band raster generated by executing the LandTrendr
+    algorithm on Google Earth Engine analyzing Landsat imagery from 1984 to
+    the user-specified year.
+
+    LandTrendr is executed twice, once on the SWIR1 band and the other on
+    Normalized Burn Ratio.
+
+    The 8-bands in returned raster are:
+        1. Years Since Disturbance (SWIR1)
+        2. Magnitude of Disturbance (SWIR1)
+        3. Duration of Disturbance (SWIR1)
+        4. Rate of Disturbance (SWIR1)
+        5. Years Since Disturbance (NBR)
+        6. Magnitude of Disturbance (NBR)
+        7. Duration of Disturbance (NBR)
+        8. Rate of Disturbance (NBR)
+
+    Parameters
+    ----------
+    bbox : list-like
+      (xmin, ymin, xmax, ymax) coordinates for bounding box
+    year : int
+      year up to which Landtrendr time series will be calculated
+    epsg : int
+      EPSG code used to define projection
+    """
+    url = get_landtrendr_download_url(bbox, year, epsg, scale)
+    ras, profile = read_gee_url(url)
+
+    return ras, profile
+
+
+def s2_from_gee(bbox, year, epsg, scale=10):
+    """Fetches an 12-band raster generated from Google Earth Engine containing
+    a leaf-off and leaf-on composite image for the user-specified area of
+    interest and year.
+
+    The 12-bands in the returned raster are:
+        1. Blue (LEAFOFF)
+        2. Green (LEAFOFF)
+        3. Red (LEAFOFF)
+        4. NIR (LEAFOFF)
+        5. SWIR1 (LEAFOFF)
+        6. SWIR2 (LEAFOFF)
+        7. Blue (LEAFON)
+        8. Green (LEAFON)
+        9. Red (LEAFON)
+        10. NIR (LEAFON)
+        11. SWIR1 (LEAFON)
+        12. SWIR2 (LEAFON)
+
+    Parameters
+    ----------
+    bbox : list-like
+      (xmin, ymin, xmax, ymax) coordinates for bounding box
+    year : int
+      year for leaf-on imagery (leaf-off will reach back into previous year)
+    epsg : int
+      EPSG code used to define projection
+    """
+    url = get_sentinel2_download_url(bbox, year, epsg, scale)
+    ras, profile = read_gee_url(url)
+
+    return ras, profile
