--- src/fetching/dem.py
@@ -1,52 +1,0 @@
-import requests
-from io import BytesIO
-from imageio import imread
-
-
-def dem_from_tnm(bbox, width, height, inSR=3857, **kwargs):
-    """
-    Retrieves a Digital Elevation Model (DEM) image from The National Map (TNM)
-    web service.
-
-    Parameters
-    ----------
-    bbox : list-like
-      list of bounding box coordinates (minx, miny, maxx, maxy)
-    width, height : int
-      desired width and height of returned image
-    inSR : int
-      spatial reference for bounding box, such as an EPSG code (e.g., 4326)
-
-    Returns
-    -------
-    dem : arr
-      DEM image as array
-    """
-    BASE_URL = ''.join([
-        'https://elevation.nationalmap.gov/arcgis/rest/',
-        'services/3DEPElevation/ImageServer/exportImage?'
-    ])
-
-    params = dict(bbox=','.join([str(x) for x in bbox]),
-                  bboxSR=inSR,
-                  size=f'{width},{height}',
-                  imageSR=inSR,
-                  time=None,
-                  format='tiff',
-                  pixelType='F32',
-                  noData=None,
-                  noDataInterpretation='esriNoDataMatchAny',
-                  interpolation='+RSP_BilinearInterpolation',
-                  compression=None,
-                  compressionQuality=None,
-                  bandIds=None,
-                  mosaicRule=None,
-                  renderingRule=None,
-                  f='image')
-    for key, value in kwargs.items():
-        params.update({key: value})
-
-    r = requests.get(BASE_URL, params=params)
-    dem = imread(BytesIO(r.content))
-
-    return dem
